+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## [Unreleased]
 
+### Added
+- Added `DriveBase.stalled()` for convenient stall detection.
+
 ### Fixed
 - Fixed drive base stall flags being set while not stalled ([support#767]).